# swagger.yaml
openapi: 3.0.0
info:
  title: CreatifyX API
  description: |
    Complete API documentation for CreatifyX - A digital asset marketplace platform.

    ## Features
    - User authentication and authorization
    - Asset management and marketplace
    - Subscription plans and payments
    - Individual asset purchases via Stripe
    - Review and rating system
    - Contact/Support ticketing system
    - Analytics dashboard for authors and admins

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ## User Roles
    - **subscriber**: Can browse, purchase assets, subscribe to plans
    - **author**: Can upload assets, view earnings analytics
    - **admin**: Can manage users, assets, plans, categories
    - **super_admin**: Full system access

  version: 1.0.0
  contact:
    name: CreatifyX API Support
    email: support@creatifyx.com
servers:
  - url: http://localhost:5000/api/v1
    description: Development server
  - url: https://api.creatifyx.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated requests
  schemas:
...
    Asset:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
          example: 'Modern Logo Pack'
        slug:
          type: string
        author:
          type: string
        assetType:
          type: string
          enum: [image, video, audio, document, '3d', template, other]
        categories:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        compatibleTools:
          type: array
          items:
            type: string
        resolution:
          type: string
          example: '1920x1080'
        orientation:
          type: string
          enum: [landscape, portrait, square]
        isPremium:
          type: boolean
        isAIGenerated:
          type: boolean
        livePreview:
          type: string
        price:
          type: number
          example: 49.99
        discountPrice:
          type: number
        status:
          type: string
          enum: [pending_review, approved, rejected]
        storage:
          type: object
          properties:
            public_id:
              type: string
            secure_url:
              type: string
            resource_type:
              type: string
            format:
              type: string
            bytes:
              type: number
        previews:
          type: object
          properties:
            thumbnail:
              type: object
            watermark:
              type: string
            images:
              type: array
              items:
                type: object
...
    Review:
      type: object
      properties:
        _id:
          type: string
        asset:
          type: string
        user:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        authorReply:
          type: string
        createdAt:
          type: string
          format: date-time
...
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        statusCode:
          type: integer
        message:
          type: string
        data:
          type: object
paths:
...
  /assets:
    post:
      tags:
        - Assets
      summary: Create asset
      description: Upload a new asset (main file + preview files)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - title
                - slug
                - assetType
                - categories
                - price
                - mainFile
              properties:
                title:
                  type: string
                slug:
                  type: string
                assetType:
                  type: string
                categories:
                  type: string
                  description: JSON array of category IDs
                tags:
                  type: string
                  description: JSON array of tags
                compatibleTools:
                  type: string
                  description: JSON array of tools
                orientation:
                  type: string
                  enum: [landscape, portrait, square]
                isPremium:
                  type: boolean
                isAIGenerated:
                  type: boolean
                price:
                  type: number
                discountPrice:
                  type: number
                mainFile:
                  type: string
                  format: binary
                preview1:
                  type: string
                  format: binary
                preview2:
                  type: string
                  format: binary
                preview3:
                  type: string
                  format: binary
                preview4:
                  type: string
                  format: binary
                preview5:
                  type: string
                  format: binary
      responses:
        '201':
          description: Asset created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Asset'

    get:
      tags:
        - Assets
      summary: Get all assets
      description: Get all assets with filtering, search, and pagination
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: assetType
          in: query
          schema:
            type: string
        - name: categories
          in: query
          schema:
            type: string
          description: Comma-separated category IDs
        - name: isPremium
          in: query
          schema:
            type: boolean
        - name: status
          in: query
          schema:
            type: string
            enum: [pending_review, approved, rejected]
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: author
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            example: '-createdAt'
      responses:
        '200':
          description: Assets retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Asset'
                      meta:
                        type: object
                        properties:
                          total:
                            type: integer
                          page:
                            type: integer
                          limit:
                            type: integer
                          totalPages:
                            type: integer
                          hasMore:
                            type: boolean
  /assets/{id}:
    get:
      tags:
        - Assets
      summary: Get asset by ID
      description: Get single asset with full details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Asset retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Asset'

    patch:
      tags:
        - Assets
      summary: Update asset
      description: Update asset status (Admin only)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending_review, approved, rejected]
      responses:
        '200':
          description: Asset updated successfully

  # ==================== SUBSCRIPTIONS ====================
...
  /reviews/asset/{assetId}:
    get:
      tags:
        - Reviews
      summary: Get reviews by asset
      description: Get all reviews for a specific asset
      parameters:
        - name: assetId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Reviews retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
                  meta:
                    type: object
...